Don't zip and preserve original paths when downloading images

+++ components/insight/SRC/org/openmicroscopy/shoola/env/data/views/calls/ArchivedImageLoader.java
@@ -64,6 +64,12 @@ public class ArchivedImageLoader
     /**Flag indicating to override or not the existing file if it exists.*/
     private boolean override;
 
+    /** Flag for zipping the downloaded images */
+    private boolean zip = true;
+    
+    /** Flag for preserving the original folder structure */
+    private boolean keepOriginalPaths = false;
+    
     /**
      * Copies the specified file to the folder.
      * 
@@ -125,19 +131,25 @@ public class ArchivedImageLoader
                 OmeroDataService os = context.getDataService();
                 File tmpFolder = null;
                 try {
-                    tmpFolder = Files.createTempDir();
+                    if(zip)
+                        tmpFolder = Files.createTempDir();
+                    else
+                        tmpFolder = folder;
                     
                     List<File> files = new ArrayList<File>();
                     
                     for (Long imageID : imageIDs) {
                         Map<Boolean, Object> r = os.getArchivedImage(ctx,
-                                tmpFolder, imageID);
+                                tmpFolder, imageID, keepOriginalPaths);
                         files.addAll((List<File>) r.get(Boolean.TRUE));
                     }
                     
                     result = new HashMap<Boolean, List<File>>();
                     
-                    if (!CollectionUtils.isEmpty(files)) {
+                    if(CollectionUtils.isEmpty(files))
+                        return;
+                    
+                    if (zip) {
                         File f = IOUtil.zipDirectory(tmpFolder, false);
                         // rename the zip
                         String baseName = FilenameUtils
@@ -150,11 +162,14 @@ public class ArchivedImageLoader
                         f = copyFile(to, folder.getParentFile());
                         ((Map<Boolean, List<File>>)result).put(Boolean.TRUE, Arrays.asList(f));
                     }
+                    else {
+                        ((Map<Boolean, List<File>>)result).put(Boolean.TRUE, files);
+                    }
                     
                 } catch (Exception e) {
                     throw new Exception(e);
                 } finally {
-                    if (tmpFolder != null)
+                    if (zip && tmpFolder != null)
                         FileUtils.deleteDirectory(tmpFolder);
                 }
             }
@@ -190,6 +205,31 @@ public class ArchivedImageLoader
     	if (CollectionUtils.isEmpty(imageIDs))
     		 throw new IllegalArgumentException("No image IDs provided.");
     	this.override = override;
+        loadCall = makeBatchCall(ctx, imageIDs, folderPath);
+    }
+
+    /**
+     * Loads the archived images.
+     * If bad arguments are passed, we throw a runtime
+     * exception so to fail early and in the caller's thread.
+     * 
+     * @param ctx The security context.
+     * @param imageID The Id of the image.
+     * @param name The name of the image.
+     * @param folderPath The location where to download the archived image.
+     * @param override Flag indicating to override the existing file if it
+     *                 exists, <code>false</code> otherwise.
+     * @param zip Pass <code>true</code> to create a zip file
+     * @param keepOriginalPaths Pass <code>true</code> to preserve the original folder structure
+     */
+    public ArchivedImageLoader(SecurityContext ctx, List<Long> imageIDs,
+            File folderPath, boolean override, boolean zip, boolean keepOriginalPaths)
+    {
+        if (CollectionUtils.isEmpty(imageIDs))
+             throw new IllegalArgumentException("No image IDs provided.");
+        this.override = override;
+        this.zip = zip;
+        this.keepOriginalPaths = keepOriginalPaths;
         loadCall = makeBatchCall(ctx, imageIDs, folderPath);
     }
 }
